Fix: Home View Cultures search logged in issue
When logged in search wasn't functioning, but now moving the search
functionality up into the Home component has resolved this issue.
