+++ src/views/Home/Cultures.tsx
@@ -1,4 +1,4 @@
-import React, { useState, useLayoutEffect } from "react";
+import React, { useState } from "react";
 import { View, FlatList } from "react-native";
 import { StackNavigationProp } from "@react-navigation/stack";
 import { ActivityIndicator, List, IconButton } from "react-native-paper";
@@ -7,7 +7,6 @@ import { Routes } from "../../routes";
 
 import { Culture, Ledger } from "../../lib";
 
-import Header from "../Header";
 import styles from "./styles";
 
 /**
@@ -18,6 +17,7 @@ type CultureProps = {
   token: string;
   cultures: Culture[];
   onRefresh: () => void;
+  searchQuery?: string;
 };
 
 /**
@@ -27,23 +27,8 @@ type CultureProps = {
  * @returns {React.ReactElement} React component
  */
 export default function Cultures(props: CultureProps): React.ReactElement {
-  const { cultures, onRefresh, token, navigation } = props;
-  const [searchQuery, setSearchQuery] = useState(undefined);
-  const [showSearch, setShowSearch] = useState(false);
+  const { cultures, onRefresh, token, searchQuery } = props;
   const [refreshing, setRefreshing] = useState(false);
-
-  useLayoutEffect(() => {
-    const header = Header({
-      title: "Cultural Awareness",
-      searchQuery: searchQuery,
-      showSearch: showSearch,
-      onSearchChange: (text: string) => setSearchQuery(text),
-      onSearchStart: () => setShowSearch(true),
-      onCancel: () => setShowSearch(false),
-    });
-
-    navigation.setOptions(header({ navigation }));
-  }, [navigation, showSearch, searchQuery]);
 
   if (!cultures) {
     return (
